Reorganize utils so future changes make more sense
