+++ src/renderers/gfmRenderer.ts
@@ -1,5 +1,4 @@
 import fs from "fs";
-import path from "path";
 
 import { Heading, toPandoc } from "@djot/djot";
 import { Environment } from "nunjucks";
@@ -11,14 +10,13 @@ import {
   DjockeyRenderer,
 } from "../types.js";
 import { runPandocOnAST } from "../pandoc.js";
-import {
-  copyFilesMatchingPattern,
-  djotASTToText,
-  ensureParentDirectoriesExist,
-  makePathBackToRoot,
-} from "../util.js";
-import { DocSet } from "../engine/docset.js";
+import { djotASTToText } from "../utils/djotUtils.js";
 import { applyFilter } from "../engine/djotFiltersPlus.js";
+import {
+  makePathBackToRoot,
+  copyFilesMatchingPattern,
+  ensureParentDirectoriesExist,
+} from "../utils/pathUtils.js";
 
 export class GFMRenderer implements DjockeyRenderer {
   identifier: DjockeyOutputFormat = "gfm";
